feat: adding showPassword hook in the login page
