+++ src/ui/pages/Onboarding/onboarding.scss
@@ -9,6 +9,11 @@
       width: 0.75rem;
     }
 
+    .eye-tooltip {
+      height: 0.75rem;
+      width: 0.75rem;
+    }
+
     .info-icon {
       color: $primary-green;
       cursor: pointer;
